add sample app IO files
This is a prototype of a way to monitor the
sample application status and set inputs without
implementing device specific functions.

The status is intended to be watched using
"watch -n 1 cat /tmp/u-phy-sample.txt"

Inputs are set by writing to /tmp/u-phy-input.txt"

An alternative way to implement device io using files
is to use a separate file for each io. That solution
may be a little cleaner but using single files for input
and status gives better overview.

Additional u-phy specific info should be added to the
status file.

Change-Id: Ibbed0e151a27b8b7b50ba45e4065ddff5865932b