add alarm support
Adds standard diagnosis support for profinet.
The states af the alarms are expected to be
handled by application. P-Net keeps track of
pending diagnosis.

The up_alarm_t should probably include strings
for generation of log messages and a status event
in the web gui.

I have not managed to get process alarms
working with gsdml. Lets use standard diagnosis
for a start and add process alarms if there is a requirement.

The diagnosis format is a bit complex for profinet.
For u-phy the functionality has been limited for easy
of use on all levels. gui/model/gsdml/test

tbd alarms showing in web gui require that alarm strings
are passed to core. The could be part of the up_alarm_t.

Change-Id: I277257cb561090be46f4342e752fcda8e2e12aa0