add configurable frame sizes
Refactors communication between core and host
so that the size of buffers used for cyclic data is
not hardcoded to 32 bytes.
The rx and tx pdo frame sizes are added to the
device configuration and will be part of code
generation.

A drawback is that the eRPC binary type is used
and every transmit will generate a heap memory
allocation.

EtherCAT adapter implementation is not updated
but still used fixed size frame buffers.

Change-Id: I035c877f2888ada86d1a1332ce544c21ff942335
Reviewed-on: https://products.intra.rt-labs.com/r/c/uniphy/up-core/+/2083
Tested-by: rtljenkins
Reviewed-by: Hans-Erik Floryd <hans-erik.floryd@rt-labs.com>