write eeprom from host
Adds api function for writing the eeprom content from the host and
updates the sample application.

The eeprom content included in the sample application is build as
an assembly file. This method of including the eeprom data is also
applied on the default eeprom data in the core EtherCAT adapter.

Refs: #2026

Change-Id: Ie4d42d4e48e74369a5832cd0e0043309b3b99e06
Reviewed-on: https://products.intra.rt-labs.com/r/c/uniphy/up-core/+/2122
Tested-by: rtljenkins
Reviewed-by: Hans-Erik Floryd <hans-erik.floryd@rt-labs.com>