add signal status to data model
Adds a status field for all signals in the model.
When a signal value is ok the status is set to UP_STATUS_OK.

The application sets the status for input signals and
the adapters will set the status for output signals.

A status area is also added to the inputs and ouputs frames. One
status byte is used for each signal. The status values are added
after the signal data area:
[ sig1_data | sig2_data | .. | sig1_status | sig2_status | .. ]

For Profinet the signal status maps to the IOxS of the subslot to
which the signal is mappped.
For EtherCAT the signal status is ok in OPERATIONAL state.

Frame sizes is not included in the device model (or generated by upgen).
Instead the frame sizes and status offsets are calculated by up-lib and core.

Handling of default values is TBD.

Change-Id: Ia570b311ac118444871fae13d6e0fa03a2407e7a