add core communication watchdog
Adds a ping rpc command which is called at a configurable
interval. If a rpc error is detected, the application is notified
using a new error callback. When an error is detected, the sample
application will reset and reconfigure the core.

To be able to restart rpc and and reconfigure the core, the
initialization procedures has been updated.

A problem with the watchdog solution is that the rpc timeout adds to
the time it takes to detect the rpc communication error. The
added error handling is useful anyway.

Change-Id: I26b6020fa21c43a4a40953aababf587aeca3db4c