add core side message queue
The message queue adds a method to send messages with
parameters from core to host. This is useful since the
core events are defined as bits and do not have parameters.

On the host side the up-lib fetches messages from the
queue one at a time for minimum impact on the timing
of the cyclic data.

Messages added:
- Profinet signal led
- Error event

A problem is that some code is duplicated for core and host side.
For example id to string functions.

Change-Id: I7b2c8515cffabc2cf6c6818f9cd13864440797bb