Fixed sample app for ECAT MONO
The sample app was crashing for EtherCAT since ESC_hw_init() wasn't
called in the monolith build. Added a call for core_main() in the sample
app to solve it which also starts the webserver which wasn't running
earlier. Also added a missing #ifdef which went missing during an
earlier merge.

Change-Id: I96272f4d7d15f8dd3bf00bcef2d7571e8484bdf1