add free running mode
In free running mode the SYNC and AVAIL events are disabled.
Instead of updating process data on these events the application
updates the data in at best effort.

Free running is the default mode of operation. The mode of operation
can be changed by the application at any time, but is natural to only
set it once, after the configuration has been written to the core.

Change-Id: I92d2ddd7fe2ee30e455196f0c92d410998897c25