Refactor sample app
The sample app is perhaps the first code seen by the user and should
be easy to understand, however currently it supports too many variants
and looks like a mess.

Refactor the code into separate implementation files for each
platform, and also separate files for the mono and client
variants. The actual application logic which is most important for the
user is in one relatively simple shared file for all variants.

Some utility functions were moved to the client and util API:s.

Change-Id: I1f8f4fcafc2e8b57fd23edee8083090d1984493a