Set free-running period to 10 ms
The previous value of 100 ms is fairly slow for most types of
devices. We should still be able to keep up with 10 ms even when using
a slow transport.

Lowering the period has a side effect in that the Python UI is updated
more often which gives a snappier feel.

Change-Id: I39c1173149e6d010a5d8b227ff688054e88cd100